Make fibonocci return a double
There is a very bad reason for this. I want to write a function that takes
the mean of a fibonocci series, but fibonocci returns an [int] and mean
expects a [double].

There are a few good solutions:

 1. I could make the math functions generic across number types.
 2. I could add automatic type casting to morloc. This is on the TODO
 list already, but I am rather far from actually implementing it in a
 general way. They user needs to have control over how much magic is in
 the air.
 3. I could add explicit casting functions.

No. 3 is probably the most practical and is also required for the
ultimate implementation of No 2.

+++ cpp_math.hpp
@@ -4,7 +4,7 @@
 double sum(std::vector<double>);
 double mean(std::vector<double>);
 double var(std::vector<double>);
-std::vector<int> fibonocci(size_t);
+std::vector<double> fibonocci(size_t);
 
 double sum(std::vector<double> xs){
     double x = 0; 
@@ -27,8 +27,8 @@ double var(std::vector<double> xs){
     return(diff / (xs.size() - 1));
 }
 
-std::vector<int> fibonocci(size_t n){
-    std::vector<int> xs;
+std::vector<double> fibonocci(size_t n){
+    std::vector<double> xs;
     if(n <= 0){
        return xs; 
     }
